Wait for MCP readiness in core integration tests
